fix compile erro

+++ 14-binary_tree_balance.c
@@ -9,14 +9,14 @@
 */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int left_height; /* Height of left subtree */
-	int right_height; /* Height of right subtree */
+	int left_height;
+	int right_height;
 
 	if (tree == NULL)
 		return (0);
 
-	left_height = binary_tree_height(tree->left);
-	right_height = binary_tree_height(tree->right);
+	left_height = (int)binary_tree_height(tree->left);
+	right_height = (int)binary_tree_height(tree->right);
 
-	return (left_height - right_height); /* Balance factor */
+	return (left_height - right_height);
 }
